requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pyyaml==6.0.2
requests==2.32.4
python-multipart==0.0.20
python-dotenv==1.1.1
